refactor: 💡 Extract functions into separate files
